refactor SDDM dominance scaling module
